+++ packages/esm-patient-queues-app/src/components/visit-form/visit-form.component.tsx
@@ -27,9 +27,9 @@ import styles from './visit-form.scss';
 import { NewVisitPayload } from '../../types';
 import { amPm, convertTime12to24 } from '../../helpers/time-helpers';
 import { useQueueRoomLocations } from '../../hooks/useQueueRooms';
-import { addQueueEntry } from '../active-visits-table.resource';
-import Overlay from '../../components/overlay/overlay.component';
-import { createVisit, useProviders } from '../patient-queues.resource';
+import { addQueueEntry } from '../../active-visits/active-visits-table.resource';
+import Overlay from '../overlay/overlay.component';
+import { createVisit, useProviders } from '../../active-visits/patient-queues.resource';
 
 interface VisitFormProps {
   patientUuid: string;
