+++ packages/esm-patient-queues-app/src/components/visit-form/visit-form.scss
