+++ packages/esm-patient-queues-app/src/index.ts
@@ -7,7 +7,7 @@ import editQueueEntryStatusModalComponent from './active-visits/change-status-di
 import patientChartMoveToNextServicePointModalComponent from './active-visits/change-status-move-to-next-dialog.component';
 import queueTableMoveToNextServicePointModalComponent from './active-visits/queue-table-move-to-next-dialog.component';
 import moveToNextServicePointButtonComponent from './active-visits/move-to-next-service-point-action.components';
-import addPatientToQueueComponent from './active-visits/visit-form/visit-form.component';
+import addPatientToQueueComponent from './components/visit-form/visit-form.component';
 import notesModalComponent from './active-visits/notes-dialog.component';
 import pickPatientEntryQueueComponent from './active-visits/pick-patient-dialog.component';
 import queueScreenComponent from './components/queue-board/queue-board.component';
@@ -17,7 +17,7 @@ import outpatientSideNavExtComponent from './side-menu/side-menu.component';
 import triageRoomComponent from './queue-triage-home.component';
 import receptionRoomComponent from './queue-reception-home.component';
 import clinicalRoomComponent from './queue-clinical-room-home.component';
-import startVisitFormComponent from './active-visits/visit-form/visit-form.component';
+import startVisitFormComponent from './components/visit-form/visit-form.component';
 import startVisitFormButtonComponent from './active-visits/start-visit-form-button.component';
 
 export const importTranslation = require.context('../translations', false, /.json$/, 'lazy');
